pyproject.toml
# SPDX-License-Identifier: LGPL-3.0-or-later

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ec_mlp"
version = "1.0"
description = "ec-MLP: Machine Learning Potential for ElectroChemistry"
readme = "README.md"
license = { file = ".license-header.txt" }
dependencies = [
    "deepmd-kit>=3.0.0",
    "torch>=2.1.0",
    "torchopt>=0.7.3",
    "numpy>=1.20.0",
    "scipy>=1.6.0",
    "pytest",
    "pytest-cov",
    "scikit_learn",
]

[project.optional-dependencies]
torch_admp = ["torch_admp @ git+https://github.com/chiahsinchu/torch-admp.git"]

[project.urls]
Homepage = "https://github.com/ChiahsinChu/ec-MLP"

[project.entry-points."deepmd"]
modifier_dipole_charge_beta_args = "ec_mlp.utils.argcheck:modifier_dipole_charge_beta_args"
modifier_dipole_charge_electrode_args = "ec_mlp.utils.argcheck:modifier_dipole_charge_electrode_args"
# modifier_dipole_charge_qeq_args = "ec_mlp.utils.argcheck:modifier_dipole_charge_qeq_args"

# [project.entry-points."deepmd.tf"]
modifier_dipole_charge_beta = "ec_mlp.tf.modifier.dipole_charge_beta:DipoleChargeBetaModifier"
modifier_dipole_charge_electrode = "ec_mlp.tf.modifier.dipole_charge_electrode:DipoleChargeElectrodeModifier"

[project.entry-points."deepmd.pt"]
modifier_dipole_charge = "ec_mlp.pt.modifier.dipole_charge:DipoleChargeModifier"

[tool.setuptools.packages.find]
where = ["src"]
include = ["ec_mlp"]
